requirements.txt
torch
numpy
PyYAML
matplotlib
tensorboard
tqdm
